Day 9

+++ Code/Day05.cs
@@ -4,7 +4,7 @@ namespace aoc2019
 {
     public class Day05
     {
-        public IEnumerable<int> Solve(string instructions, int input)
+        public IEnumerable<long> Solve(string instructions, int input)
         {
             return new IntCodeComputer(instructions)
                 .AddInput(input)
